crud

+++ main/models.py
@@ -6,7 +6,8 @@ class Product(models.Model):
     name = models.CharField(max_length=100)
     description = models.TextField()
     product_image = models.ImageField(upload_to='products/%Y/%m/%d/', blank=True, null=True)
-    price = models.IntegerField()
+    price = models.DecimalField(max_digits=10, decimal_places=2)
+
     
 
     def __str__(self):
